Split off claims package from storage
