+++ pkg/claims/doc.go
@@ -1,0 +1,1 @@
+package claims // import "get.porter.sh/porter/pkg/claims"
