+++ pkg/cnab/provider/helpers.go
@@ -1,10 +1,11 @@
 package cnabprovider
 
 import (
+	"testing"
+
+	"get.porter.sh/porter/pkg/claims"
 	"get.porter.sh/porter/pkg/config"
 	"get.porter.sh/porter/pkg/credentials"
-	"get.porter.sh/porter/pkg/storage"
-	"testing"
 )
 
 type TestRuntime struct {
@@ -14,11 +15,11 @@ type TestRuntime struct {
 
 func NewTestRuntime(t *testing.T) *TestRuntime {
 	c := config.NewTestConfig(t)
-	claimStorage := storage.NewTestClaimProvider()
+	claimStorage := claims.NewTestClaimProvider()
 	credentialStorage := credentials.NewTestCredentialProvider(t, c)
 
 	return &TestRuntime{
-		TestConfig:   c,
-		Runtime: NewRuntime(c.Config, claimStorage, credentialStorage),
+		TestConfig: c,
+		Runtime:    NewRuntime(c.Config, claimStorage, credentialStorage),
 	}
 }
