+++ pkg/storage/crudstore/plugin.go
@@ -7,14 +7,14 @@ import (
 	"github.com/hashicorp/go-plugin"
 )
 
-// PluginInterface for the claim storage. This first part of the
+// PluginInterface for the data storage. This first part of the
 // three-part plugin key is only seen/used by the plugins when the host is
 // communicating with the plugin and is not exposed to users.
 const PluginInterface = "storage"
 
 var _ plugin.Plugin = &Plugin{}
 
-// Plugin is a generic type of plugin for working with any implementation of a claim store.
+// Plugin is a generic type of plugin for working with any implementation of a crud store.
 type Plugin struct {
 	Impl crud.Store
 }
